requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.20
langchain-openai==0.1.6
langchain-community==0.0.38
langchain-core>=0.1.52
chromadb==0.4.15
pandas==2.1.3
numpy==1.26.2
duckdb==0.9.2
matplotlib==3.8.2
seaborn==0.13.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

